Initial check-in.
This is the initial check-in of the Excel demo, such as it is.  At present, the debugger is of extremely limited use-- breakpoints cannot be set, and a number of error conditions will apparently clobber previous console output before it gets back to Visual Studio, which can be very misleading.  Additionally, there is a lot of material in here that Visual Studio threw into the original project; it's not clear to me that we can safely delete it until we figure out what would re-add it.  See the README file for further details.